MDEV-12215: main.repair_symlink-5543 fails in buildbot
If openat is present on the system and it tries to open a symlink with
O_NOFOLLOW, we get errno 40.

If openat is not present on the system, we use the alternative open call, with
slightly different logic. IF the symlink doesn't point to a valid file,
we get errno 20. This test uses an invalid symlink on the table t1.MYD.