Fix issues caught by Clang's -Wpointer-bool-conversion warning
Fix instances where a pointer is always converted to a true value. Since
this value is known beforehand, it can be safely removed. These include
checking the address of an array, taking the address of a local variable, or
taking the address of a member variable. In most cases, the pointer is removed
from the conditional. In a few instances, the address-of operator is removed,
or the check is performed on the first element of the array. In a couple
of cases, attribute(nonnull) is changed in function declarations.

This only applies to the core client/server (and InnoDB), other plugins or storage
engines may still have issues.