Porting the proxy protocol code from the latest percona
The proxy protocol allows a proxy front-end to communicate
the client's IP address to a backend server. See [1,2]
for more information.

The code was ported and adapted from the current Percona
git base, using the original commit [3] as a guide
for identifying the files. No major change was applied
to any of the code at this time.

[1] https://blog.haproxy.com/2017/03/24/using-haproxy-with
-the-proxy-protocol-to-better-secure-your-database/
[2] http://www.haproxy.org/download/1.8/doc/proxy-protocol.txt
[3] https://github.com/percona/percona-server/commit/
c464c4de5dc8d07576e154da70745eab8d2b392f