Do not crash XtraDB when fil_space_acquire() fails
This reverts part of commit 50eb40a2a8aa3af6cc271f6028f4d6d74301d030
which backported the code from MariaDB 10.2. The XtraDB version of
the code included a ut_error statement (aborting the process) when
a tablespace is not found. Luckily this change was not part of a
release; MariaDB 10.1.22 had been released some days earlier.