MDEV-12179: Per-engine mysql.gtid_slave_pos table
Intermediate commit.

Fix test failure in rpl.rpl_gtid_mdev4484 when an earlier test changed
the mysql.gtid_slave_pos table to innodb. With MDEV-12179, the delete
of the old GTID is skipped, because it is recorded as belonging to
another engine, due to ALTER TABLE in the test case.

Fix by restart the SQL thread after the ALTER to have the correct
engine registered.

Also fix the test case rpl.rpl_gtid_ignored to not leave
mysql.innodb_slave_pos as a different engine.